Before you Begin:
Create a new repository for this project called python-challenge.

Clone the new repository to your computer.

Inside your local git repository, create a directory for both of the  Python Challenges. Use folder names corresponding to the challenges: PyBank and  PyPoll.

Inside of each folder that you just created, add the following:
-- A new file called main.py. This will be the main script to run for each analysis.
-- A "Resources" folder that contains the CSV files you used. Make sure your script has the correct path to the CSV file.
-- An "analysis" folder that contains your text file that has the results from your analysis.

Push the above changes to GitHub or GitLab.